[Continuous batching] Late token vector initialization in sampling (#649)
Changes:
- Further split of greedy and multinomial paths - using original logits
buffer in greedy and whenever possible in multinomial sampling. Sorted
vector is created only when top_p or top_k filters need to be applied.
- Fixing issue with top_k filter being applied always when multinomial
sampling is used unless it's explicitly set to 0. Now default value
(which is max for size_t) will not trigger applying top_k filter. The
filter will also not be applied if top_k is bigger than logits vector
size.
- Skipping multinomial tests